chore: reorganize and update integration tests for file uploads and user retrieval
