+++ notionrs/tests/mutable/page/mod.rs
@@ -1,4 +1,3 @@
 mod create_page_on_data_source;
 mod crud_page;
-mod get_page;
 mod move_page;
