+++ notionrs/tests/readonly/search.rs
@@ -9,9 +9,10 @@ mod integration_tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn search() -> Result<(), notionrs::Error> {
-        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
+        dotenvy::from_path(std::path::Path::new(".env.readonly"))
+            .expect("Failed to load .env.readonly file");
 
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let request = client.search().query("").sort_timestamp_asc();
@@ -32,9 +33,9 @@ mod integration_tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn search_page() -> Result<(), notionrs::Error> {
-        dotenvy::from_path(std::path::Path::new("../.env")).ok();
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let request = client.search_page().query("").sort_timestamp_asc();
@@ -55,9 +56,9 @@ mod integration_tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn search_database() -> Result<(), notionrs::Error> {
-        dotenvy::from_path(std::path::Path::new("../.env")).ok();
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let request = client.search_database().query("").sort_timestamp_asc();
