+++ notionrs/tests/readonly/user/get_user.rs
@@ -2,17 +2,15 @@ mod integration_tests {
 
     #[tokio::test]
     async fn get_user() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
+        dotenvy::from_path(std::path::Path::new(".env.readonly"))
+            .expect("Failed to load .env.readonly file");
 
-                let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let res = client.get_self().send().await?;
 
         let user_id = res.id;
-
-                let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
-        let client = notionrs::Client::new(notion_api_key);
 
         let request = client.get_user().user_id(user_id);
